check for overwrite

+++ composer/utils/checkpoint.py
@@ -414,6 +414,7 @@ def save_checkpoint(
     filename: str = 'ep{epoch}-ba{batch}-rank{rank}',
     *,
     weights_only: bool = False,
+    overwrite: bool = False,
 ) -> Union[str, None]:  # noqa: D103
 
     log.debug('Saving checkpoint to %s', filename)
@@ -433,9 +434,12 @@ def save_checkpoint(
     if dirname:
         os.makedirs(dirname, exist_ok=True)
 
+    if os.path.exists(save_filename) and not overwrite:
+        raise FileExistsError(f'{save_filename}. To overwrite, set overwrite=True.')
+
     # only rank 0 saves the state_dict
     if dist.get_global_rank() == 0:
-        with open(save_filename, 'xb') as f:
+        with open(save_filename, 'wb') as f:
             torch.save(state_dict, f)
 
         if is_tar(save_filename):
